+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 - Add clean shutdown hook for IdentifierMapper to clean up the worker threads
+- Added new application for reviewing IsIdentifiable output files
 
 ## [1.4.5] - 2020-02-26